Added sensor communication branch changes manually
