+++ src/views/Feed.vue
@@ -34,12 +34,8 @@ Register.vue - base vue
 <script>
 import NavBarTop from "../components/navbars/NavBarTop.vue";
 import { XsensDotSensor } from "/src/libraries/bluetooth.js";
-import {
-  findBluetoothDevices,
-  startRTStream,
-  stopRTStream,
-  syncSensor,
-} from "/src/libraries/bluetooth.js";
+import { findBluetoothDevices, startRTStream, stopRTStream, getSyncStatusSensor } from "/src/libraries/interface.js";
+
 export default {
   name: "Feed",
   components: {
